+++ pkg/workspace_template.go
@@ -22,13 +22,165 @@ import (
 	"strings"
 )
 
+// createWorkspaceTemplateVersionDB creates a workspace template version in the database.
+func createWorkspaceTemplateVersionDB(tx *sql.Tx, workspaceTemplateID uint64, version int64, manifest string, isLatest bool) (id uint64, err error) {
+	err = sb.Insert("workspace_template_versions").
+		SetMap(sq.Eq{
+			"version":               version,
+			"is_latest":             isLatest,
+			"manifest":              manifest,
+			"workspace_template_id": workspaceTemplateID,
+		}).
+		Suffix("RETURNING id").
+		RunWith(tx).
+		QueryRow().
+		Scan(&id)
+
+	return
+}
+
+// markWorkspaceTemplateVersionsOutdatedDB updates all of the workspace template versions in db so is_latest is false
+// given the workspaceTemplateID
+func markWorkspaceTemplateVersionsOutdatedDB(tx *sql.Tx, workspaceTemplateID uint64) (err error) {
+	_, err = sb.Update("workspace_template_versions").
+		SetMap(sq.Eq{"is_latest": false}).
+		Where(sq.Eq{
+			"workspace_template_id": workspaceTemplateID,
+			"is_latest":             true,
+		}).
+		RunWith(tx).
+		Exec()
+
+	return
+}
+
+// createLatestWorkspaceTemplateVersionDB creates a new workspace template version and marks all previous versions as not latest.
+func createLatestWorkspaceTemplateVersionDB(tx *sql.Tx, workspaceTemplateID uint64, version int64, manifest string) (id uint64, err error) {
+	id, err = createWorkspaceTemplateVersionDB(tx, workspaceTemplateID, version, manifest, true)
+	if err != nil {
+		return
+	}
+
+	err = markWorkspaceTemplateVersionsOutdatedDB(tx, workspaceTemplateID)
+
+	return
+}
+
 func parseWorkspaceSpec(template string) (spec *WorkspaceSpec, err error) {
 	err = yaml.UnmarshalStrict([]byte(template), &spec)
 
 	return
 }
 
-func generateArguments(spec *WorkspaceSpec, config map[string]string, withRuntimeVars bool) (err error) {
+func generateRuntimeParamters(config SystemConfig) (parameters []Parameter, err error) {
+	parameters = make([]Parameter, 0)
+
+	// Host
+	parameters = append(parameters, Parameter{
+		Name:  "sys-host",
+		Value: config.Domain(),
+		Type:  "input.hidden",
+	})
+
+	// Node pool parameter and options
+	options, err := config.ParsedNodePoolOptions()
+	if err != nil {
+		return nil, err
+	}
+	if len(options) == 0 {
+		return nil, fmt.Errorf("no node pool options in config")
+	}
+
+	parameters = append(parameters, Parameter{
+		Name:        "sys-node-pool",
+		Value:       ptr.String(options[0].Value),
+		Type:        "select.select",
+		Options:     options,
+		DisplayName: ptr.String("Node pool"),
+		Hint:        ptr.String("Name of node pool or group"),
+		Required:    true,
+	})
+
+	return
+}
+
+func generateStaticParameters() (parameters []Parameter, err error) {
+	parameters = make([]Parameter, 0)
+
+	// Resource action parameter
+	parameters = append(parameters, Parameter{
+		Name:        "sys-name",
+		Type:        "input.text",
+		Value:       ptr.String("name"),
+		DisplayName: ptr.String("Workspace name"),
+		Hint:        ptr.String("Must be between 3-30 characters, contain only alphanumeric or `-` characters"),
+		Required:    true,
+	})
+
+	// TODO: These can be removed when lint validation of workflows work
+	// Resource action parameter
+	parameters = append(parameters, Parameter{
+		Name:  "sys-resource-action",
+		Value: ptr.String("apply"),
+		Type:  "input.hidden",
+	})
+	// Workspace action
+	parameters = append(parameters, Parameter{
+		Name:  "sys-workspace-action",
+		Value: ptr.String("create"),
+		Type:  "input.hidden",
+	})
+
+	// UID placeholder
+	parameters = append(parameters, Parameter{
+		Name:  "sys-uid",
+		Value: ptr.String("uid"),
+		Type:  "input.hidden",
+	})
+
+	return
+}
+
+func generateVolumeParameters(spec *WorkspaceSpec) (parameters []Parameter, err error) {
+	if spec == nil {
+		return nil, fmt.Errorf("workspaceSpec is nil")
+	}
+
+	parameters = make([]Parameter, 0)
+
+	// Map all the volumeClaimTemplates that have storage set
+	volumeStorageQuantityIsSet := make(map[string]bool)
+	for _, v := range spec.VolumeClaimTemplates {
+		if v.Spec.Resources.Requests != nil {
+			volumeStorageQuantityIsSet[v.ObjectMeta.Name] = true
+		}
+	}
+	// Volume size parameters
+	volumeClaimsMapped := make(map[string]bool)
+	for _, c := range spec.Containers {
+		for _, v := range c.VolumeMounts {
+			// Skip if already mapped or storage size is set
+			if volumeClaimsMapped[v.Name] || volumeStorageQuantityIsSet[v.Name] {
+				continue
+			}
+
+			parameters = append(parameters, Parameter{
+				Name:        fmt.Sprintf("sys-%v-volume-size", v.Name),
+				Type:        "input.number",
+				Value:       ptr.String("20480"),
+				DisplayName: ptr.String(fmt.Sprintf("Disk size for \"%v\"", v.Name)),
+				Hint:        ptr.String(fmt.Sprintf("Disk size in MB for volume mounted at `%v`", v.MountPath)),
+				Required:    true,
+			})
+
+			volumeClaimsMapped[v.Name] = true
+		}
+	}
+
+	return
+}
+
+func generateArguments(spec *WorkspaceSpec, config SystemConfig, withRuntimeVars bool) (err error) {
 	systemParameters := make([]Parameter, 0)
 	// Resource action parameter
 	systemParameters = append(systemParameters, Parameter{
@@ -65,7 +217,7 @@ func generateArguments(spec *WorkspaceSpec, config map[string]string, withRuntim
 		// Host
 		systemParameters = append(systemParameters, Parameter{
 			Name:  "sys-host",
-			Value: ptr.String(config["ONEPANEL_DOMAIN"]),
+			Value: config.Domain(),
 			Type:  "input.hidden",
 		})
 
@@ -590,12 +742,6 @@ func (c *Client) createWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 	}
 	workspaceTemplate.UID = uid
 
-	tx, err := c.DB.Begin()
-	if err != nil {
-		return nil, err
-	}
-	defer tx.Rollback()
-
 	workspaceTemplate.WorkflowTemplate.IsSystem = true
 	workspaceTemplate.WorkflowTemplate.Resource = ptr.String(TypeWorkspaceTemplate)
 	workspaceTemplate.WorkflowTemplate.ResourceUID = ptr.String(uid)
@@ -606,6 +752,11 @@ func (c *Client) createWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 	workspaceTemplate.Version = workspaceTemplate.WorkflowTemplate.Version
 	workspaceTemplate.IsLatest = true
 
+	tx, err := c.DB.Begin()
+	if err != nil {
+		return nil, err
+	}
+	defer tx.Rollback()
 	err = sb.Insert("workspace_templates").
 		SetMap(sq.Eq{
 			"uid":                  uid,
@@ -615,7 +766,8 @@ func (c *Client) createWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 		}).
 		Suffix("RETURNING id, created_at").
 		RunWith(tx).
-		QueryRow().Scan(&workspaceTemplate.ID, &workspaceTemplate.CreatedAt)
+		QueryRow().
+		Scan(&workspaceTemplate.ID, &workspaceTemplate.CreatedAt)
 	if err != nil {
 		_, errCleanUp := c.ArchiveWorkflowTemplate(namespace, workspaceTemplate.WorkflowTemplate.UID)
 		errorMsg := "Error with insert into workspace_templates. "
@@ -626,39 +778,27 @@ func (c *Client) createWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 		return nil, util.NewUserErrorWrap(err, errorMsg) //return the source error
 	}
 
-	workspaceTemplateVersionID := uint64(0)
-	err = sb.Insert("workspace_template_versions").
-		SetMap(sq.Eq{
-			"version":               workspaceTemplate.Version,
-			"is_latest":             workspaceTemplate.IsLatest,
-			"manifest":              workspaceTemplate.Manifest,
-			"workspace_template_id": workspaceTemplate.ID,
-		}).
-		Suffix("RETURNING id").
-		RunWith(tx).
-		QueryRow().
-		Scan(&workspaceTemplateVersionID)
+	workspaceTemplateVersionID, err := createWorkspaceTemplateVersionDB(tx, workspaceTemplate.ID, workspaceTemplate.Version, workspaceTemplate.Manifest, true)
 	if err != nil {
-		_, errCleanUp := c.ArchiveWorkflowTemplate(namespace, workspaceTemplate.WorkflowTemplate.UID)
 		errorMsg := "Error with insert into workspace_templates_versions. "
+		_, errCleanUp := c.ArchiveWorkflowTemplate(namespace, workspaceTemplate.WorkflowTemplate.UID)
 		if errCleanUp != nil {
+			err = fmt.Errorf("%w; %s", err, errCleanUp)
 			errorMsg += "Error with clean-up: ArchiveWorkflowTemplate. "
-			errorMsg += errCleanUp.Error()
 		}
-		return nil, util.NewUserErrorWrap(err, errorMsg) //return the source error
+		return nil, util.NewUserErrorWrap(err, errorMsg) // return the source error
 	}
 
-	if len(workspaceTemplate.Labels) != 0 {
-		_, err = c.InsertLabelsBuilder(TypeWorkspaceTemplateVersion, workspaceTemplateVersionID, workspaceTemplate.Labels).
-			RunWith(tx).
-			Exec()
-		if err != nil {
-			return nil, err
-		}
+	_, err = c.InsertLabelsRunner(tx, TypeWorkspaceTemplateVersion, workspaceTemplateVersionID, workspaceTemplate.Labels)
+	if err != nil {
+		return nil, err
 	}
 
 	if err = tx.Commit(); err != nil {
-		_, err := c.ArchiveWorkflowTemplate(namespace, workspaceTemplate.WorkflowTemplate.UID)
+		_, errArchive := c.ArchiveWorkflowTemplate(namespace, workspaceTemplate.WorkflowTemplate.UID)
+		if errArchive != nil {
+			err = fmt.Errorf("%w; %s", err, errArchive)
+		}
 		return nil, err
 	}
 
@@ -666,7 +806,7 @@ func (c *Client) createWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 }
 
 func (c *Client) workspaceTemplatesSelectBuilder(namespace string) sq.SelectBuilder {
-	sb := sb.Select(getWorkspaceTemplateColumns("wt", "")...).
+	sb := sb.Select(getWorkspaceTemplateColumns("wt")...).
 		From("workspace_templates wt").
 		Where(sq.Eq{
 			"wt.namespace": namespace,
@@ -775,8 +915,8 @@ func (c *Client) GenerateWorkspaceTemplateWorkflowTemplate(workspaceTemplate *Wo
 
 // CreateWorkspaceTemplate creates a template for Workspaces
 func (c *Client) CreateWorkspaceTemplate(namespace string, workspaceTemplate *WorkspaceTemplate) (*WorkspaceTemplate, error) {
-	valid, err := govalidator.ValidateStruct(workspaceTemplate)
-	if err != nil || !valid {
+	_, err := govalidator.ValidateStruct(workspaceTemplate)
+	if err != nil {
 		return nil, util.NewUserError(codes.InvalidArgument, err.Error())
 	}
 
@@ -784,6 +924,7 @@ func (c *Client) CreateWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 	if err != nil {
 		return nil, err
 	}
+
 	if existingWorkspaceTemplate != nil {
 		message := fmt.Sprintf("Workspace template with the name '%v' already exists", workspaceTemplate.Name)
 		if existingWorkspaceTemplate.IsArchived {
@@ -864,55 +1005,30 @@ func (c *Client) UpdateWorkspaceTemplate(namespace string, workspaceTemplate *Wo
 	updatedWorkflowTemplate.ID = existingWorkspaceTemplate.WorkflowTemplate.ID
 	updatedWorkflowTemplate.UID = existingWorkspaceTemplate.WorkflowTemplate.UID
 
-	tx, err := c.DB.Begin()
-	if err != nil {
-		return nil, err
-	}
-	defer tx.Rollback()
-
 	updatedWorkflowTemplate.Labels = workspaceTemplate.Labels
 	workflowTemplateVersion, err := c.CreateWorkflowTemplateVersion(namespace, updatedWorkflowTemplate)
 	if err != nil {
 		return nil, err
 	}
 
+	// TODO - this might not be needed with recent changes made.
 	workspaceTemplate.Version = workflowTemplateVersion.Version
 	workspaceTemplate.IsLatest = true
 
-	_, err = sb.Update("workspace_template_versions").
-		SetMap(sq.Eq{"is_latest": false}).
-		Where(sq.Eq{
-			"workspace_template_id": workspaceTemplate.ID,
-		}).
-		RunWith(tx).
-		Exec()
+	tx, err := c.DB.Begin()
+	if err != nil {
+		return nil, err
+	}
+	defer tx.Rollback()
+
+	workspaceTemplateVersionID, err := createLatestWorkspaceTemplateVersionDB(tx, workspaceTemplate.ID, workspaceTemplate.Version, workspaceTemplate.Manifest)
 	if err != nil {
 		return nil, err
 	}
 
-	workspaceTemplateVersionID := uint64(0)
-	err = sb.Insert("workspace_template_versions").
-		SetMap(sq.Eq{
-			"version":               workspaceTemplate.Version,
-			"is_latest":             workspaceTemplate.IsLatest,
-			"manifest":              workspaceTemplate.Manifest,
-			"workspace_template_id": workspaceTemplate.ID,
-		}).
-		Suffix("RETURNING id").
-		RunWith(tx).
-		QueryRow().
-		Scan(&workspaceTemplateVersionID)
+	_, err = c.InsertLabelsRunner(tx, TypeWorkspaceTemplateVersion, workspaceTemplateVersionID, workspaceTemplate.Labels)
 	if err != nil {
 		return nil, err
-	}
-
-	if len(workspaceTemplate.Labels) != 0 {
-		_, err = c.InsertLabelsBuilder(TypeWorkspaceTemplateVersion, workspaceTemplateVersionID, workspaceTemplate.Labels).
-			RunWith(tx).
-			Exec()
-		if err != nil {
-			return nil, err
-		}
 	}
 
 	if err := tx.Commit(); err != nil {
@@ -958,9 +1074,7 @@ func (c *Client) ListWorkspaceTemplateVersions(namespace, uid string) (workspace
 		return
 	}
 
-	ids := WorkspaceTemplatesToVersionIds(workspaceTemplates)
-
-	labelsMap, err := c.GetDbLabelsMapped(TypeWorkspaceTemplateVersion, ids...)
+	labelsMap, err := c.GetDBLabelsMapped(TypeWorkspaceTemplateVersion, WorkspaceTemplatesToVersionIDs(workspaceTemplates)...)
 	if err != nil {
 		return nil, err
 	}
