+++ pkg/workspace_template_types.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	"github.com/onepanelio/core/pkg/util/ptr"
+	"github.com/onepanelio/core/sqlutil"
 	"gopkg.in/yaml.v2"
 	"time"
 )
@@ -175,7 +176,7 @@ func (wt *WorkspaceTemplate) InjectRuntimeVariables(config SystemConfig) error {
 	return nil
 }
 
-func WorkspaceTemplatesToVersionIds(resources []*WorkspaceTemplate) (ids []uint64) {
+func WorkspaceTemplatesToVersionIDs(resources []*WorkspaceTemplate) (ids []uint64) {
 	mappedIds := make(map[uint64]bool)
 
 	// This is to make sure we don't have duplicates
@@ -192,7 +193,7 @@ func WorkspaceTemplatesToVersionIds(resources []*WorkspaceTemplate) (ids []uint6
 
 // getWorkspaceTemplateColumns returns all of the columns for workspace template modified by alias, destination.
 // see formatColumnSelect
-func getWorkspaceTemplateColumns(alias string, destination string, extraColumns ...string) []string {
+func getWorkspaceTemplateColumns(aliasAndDestination ...string) []string {
 	columns := []string{"id", "uid", "created_at", "modified_at", "name", "namespace", "is_archived", "workflow_template_id"}
-	return formatColumnSelect(columns, alias, destination, extraColumns...)
+	return sqlutil.FormatColumnSelect(columns, aliasAndDestination...)
 }
